chapters/4-Images/pyproject.toml
[tool.poetry]
name = "mandybrot"
version = "0.0.0"
description = "Plot the magical Mandelbrot set."
authors = ["FreddyWordingham <freddy@digilab.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.24.2"
pillow = "^9.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
